2019 day 10 part 2
This was a bitch... First I was just using the slope between the asteroids.
That kind of worked, but lead to some interesting edge cases where slope
was infinite or 0.  Hard to distinguish up from down or left from right
in those cases.  So I moved on to radians but then started counting
asteroids at the wrong place (i.e. figuring out "up" for where to start
So I moved to using degrees and that helped smooth things out a lot.
Also haviing to use the reverse angle became really important.